requirements.txt
Flask
tensorflow
Redis
Pillow
cassandra-driver
